mmv updates

+++ .oh-my-zsh-custom/functions.zsh
@@ -1,3 +1,7 @@
+lfcd () {
+    'cd' "$('lf' -print-last-dir "$@")"
+}
+
 # portfoward() {
 #   if [ $# -lt 2 ]
 #   then
@@ -10,34 +14,42 @@
 
 matrix() { echo -e "\e[1;40m" ; clear ; while :; do echo $LINES $COLUMNS $(( $RANDOM % $COLUMNS)) $(( $RANDOM % 72 )) ;sleep 0.05; done|awk '{ letters="abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789@#$%^&*()"; c=$4;        letter=substr(letters,c,1);a[$3]=0;for (x in a) {o=a[x];a[x]=a[x]+1; printf "\033[%s;%sH\033[2;32m%s",o,x,letter; printf "\033[%s;%sH\033[1;37m%s\033[0;0H",a[x],x,letter;if (a[x] >= $1) { a[x]=0; } }}' }
 
-mmv ()
+mmv()
 {
+    OLD=/tmp/current_files
+    NEW=/tmp/rename_files
+
+    # get options
     dry='false'
     verbose='false'
-
-    print_usage() {
-        printf "Usage: ..."
-    }
-
-    while getopts 'dv' flag; do
+    while getopts 'dvh' flag; do
         case "${flag}" in
             d) dry='true' ;;
             v) verbose='true' ;;
-            *) print_usage
-               exit 1 ;;
+            h | *) echo "Usage: $0 [-d  dryrun] [-v  verbose] [file or folder]" 1>&2
+                   return 0;;
         esac
     done
+
     if "$dry"; then
+        echo "Dry run:"
         cmd=echo
+        NEW=${NEW}_dry_run
     else
         cmd=''
     fi
 
+    # get the directory argument
+    DIR="${@: -1}"
+    if [ $DIR ] && [ -d $DIR ]; then
+        DIR="${DIR%/}/"
+    else
+        DIR=""
+    fi
+    echo "DIR: $DIR"
 
-    OLD=/tmp/current_files
-    NEW=/tmp/renamed_files
-
-    ls -1 > $OLD
+    # get the files and open the editor
+    ls -1p $DIR | sed "s|^|$DIR|" > $OLD
     cp $OLD $NEW
     ${EDITOR:-vi} $NEW
 
@@ -53,18 +65,26 @@ mmv ()
         return 1
     fi
 
-    count=0
+    # rename or delete the files
+    mv_count=0
+    rm_count=0
     while IFS= read -r orig_fname && IFS= read -r new_fname <&3; do
         if [ -z "$new_fname" ]; then
             $cmd rm "$orig_fname"
-            ((count+=1))
+            ((rm_count+=1))
         elif [ "$orig_fname" != "$new_fname" ]; then
-            $cmd mv "$orig_fname" "$new_fname"
-            ((count+=1))
+            # check if new_fname exists and don't overwrite
+            if [ -e "$new_fname" ]; then
+                echo "Cannot mv $orig_fname $new_fname, new filename already exists. Skipping"
+            else
+                $cmd mv "$orig_fname" "$new_fname"
+                ((mv_count+=1))
+            fi
         fi
     done < $OLD 3< $NEW
 
-    echo "Renamed $count files"
+    [ $mv_count -gt 0 ] && echo "Renamed $mv_count files"
+    [ $rm_count -gt 0 ] && echo "Removed $rm_count files"
 
     rm $OLD $NEW
     return 0
